backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.2",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "openai>=2.9.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "uvicorn>=0.38.0",
    "json_repair>=0.35.0",
    "toml>=0.10.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
